+++ src/app/components/order-history/order-history.module.ts
@@ -1,0 +1,38 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatButtonModule } from '@angular/material/button';
+import { MatChipsModule } from '@angular/material/chips';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatTooltipModule } from '@angular/material/tooltip';
+
+import { MessageService } from '../../services/message/message.module';
+
+import { OrderHistoryComponent } from './order-history.component';
+
+export { OrderHistoryComponent } from './order-history.component';
+
+@NgModule({
+  declarations: [
+    OrderHistoryComponent
+  ],
+  exports: [
+    OrderHistoryComponent
+  ],
+  imports: [
+    CommonModule,
+    FlexLayoutModule,
+    MatButtonModule,
+    MatChipsModule,
+    MatIconModule,
+    MatInputModule,
+    MatSlideToggleModule,
+    MatTooltipModule
+  ],
+  providers: [
+    MessageService
+  ]
+})
+export class OrderHistoryModule { }