+++ src/app/core/add-on.ts
@@ -1,4 +1,5 @@
 export class AddOn {
+  id: number;
   name: string;
   type: string;