+++ src/app/core/user-info.ts
@@ -1,3 +1,4 @@
+import { Coordinate } from './coordinate';
 import { DeliveryLocation } from './delivery-location';
 
 export class UserInfo {
@@ -7,7 +8,9 @@ export class UserInfo {
   adminEnabled: boolean;
   defaultView: string;
   defaultDeliveryLocation: DeliveryLocation;
+  pickupLocation: Coordinate;
   locale: string;
+  pictureUrl: string;
 
   constructor(that: UserInfo) {
     this.username = that.username;
@@ -16,7 +19,9 @@ export class UserInfo {
     this.adminEnabled = that.adminEnabled;
     this.defaultView = that.defaultView;
     this.defaultDeliveryLocation = DeliveryLocation.copy(that.defaultDeliveryLocation);
+    this.pickupLocation = Coordinate.copy(that.pickupLocation);
     this.locale = that.locale;
+    this.pictureUrl = that.pictureUrl;
   }
 
 }