+++ src/app/services/message/message.service.ts
@@ -1,116 +1,52 @@
-// this is required since "Observable" doesn't include interval on import
-import 'rxjs/add/observable/interval';
-
-import * as _ from 'lodash';
-
 import { HttpClient } from '@angular/common/http';
-import { Inject, Injectable, OnDestroy } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Inject, Injectable } from '@angular/core';
 import { Subject } from 'rxjs/Subject';
-import { Subscription } from 'rxjs/Subscription';
 
-import { AppConfig } from '../../app.config';
+import { environment } from '../../../environments/environment';
+
 import { Batch } from '../../core/batch';
 import { DeliveryLocation } from '../../core/delivery-location';
 import { EnvironmentDetails } from '../../core/environment-details';
-import { isDefined } from '../../core/is-defined';
 import { Order } from '../../core/order';
 import { OrderInfo } from '../../core/order-info';
 import { OrderOption } from '../../core/order-option';
-import { RobotStatus } from '../../core/robot-status';
+import { SystemDetails } from '../../core/system-details';
 import { UserInfo } from '../../core/user-info';
 
 @Injectable()
-export class MessageService implements OnDestroy {
-
-  private static BATCH_POLLING_RATE = 250;
-  private static ORDER_POLLING_RATE = 250;
-  private static DELIVERED = 'Delivered';
-  private static WAITING = 'Waiting';
-  private static DELIVERING = 'Delivering';
-  private static DELAYED = 'Delayed';
+export class MessageService {
 
   private sessionKey: string;
-  private user: UserInfo;
-  private environmentDetails: EnvironmentDetails;
+  private user: UserInfo; // todo: would be nice to get rid of this, don't think it's possible
 
-  private orderUpdatesTimer: Observable<number>;
-  private orderUpdatesSubscription: Subscription;
-  private orderHistoryCache: Order[] = [];
-  private ordersToMonitor: Order[] = [];
-  private orderHistorySubject: Subject<Order[]>;
+  userUpdates: Subject<UserInfo> = new Subject<UserInfo>();
+  orderPlacedUpdate: Subject<Order> = new Subject<Order>();
 
-  private batchUpdatesTimer: Observable<number>;
-  private batchUpdatesSubscription: Subscription;
-  private batchUpdatesSubject: Subject<Batch[]>;
-  private batchUpdatesCache: Batch[] = [];
+  constructor(private http: HttpClient) {}
 
-  constructor(@Inject(AppConfig) private config: AppConfig, private http: HttpClient) {}
-
-  private fetchEnvironmentDetails(): Promise<EnvironmentDetails> {
+  getEnvironmentDetails(): Promise<EnvironmentDetails> {
     return new Promise<any>((resolve, reject) => {
       const body = {
         username: this.user.username,
         sessionKey: this.sessionKey
       };
 
-      this.http.post(this.config.backendUrl + 'getEnvironmentDetails', body)
+      this.http.post(environment.backendUrl + 'getEnvironmentDetails', body)
         .subscribe((response: EnvironmentDetails) => resolve(response), error => reject(error));
     });
   }
 
-  private handleGetOrderUpdates(): void {
-    // this is when the timer has triggered, need to fetch orders from the monitored Orders
-    // then merge that with the cache (updating where needed)
-    // emit the new cache to the subject
-
-    const body = {
-      username: this.user.username,
-      sessionKey: this.sessionKey,
-      orders: this.ordersToMonitor
-    };
-    this.http.post(this.config.backendUrl + 'getOrderUpdates', body)
-      .subscribe(
-        (response: {orders: Order[]}) => {
-          response.orders.forEach((order: Order) => {
-            const i = _.findIndex(this.orderHistoryCache, (cachedOrder: Order) => {
-              return cachedOrder.id === order.id;
-            });
-
-            i === -1 ? this.orderHistoryCache.push(order) : this.orderHistoryCache[i] = order;
-
-            if (order.state === MessageService.DELIVERED) {
-              const j = _.findIndex(this.ordersToMonitor, (monitoredOrder: Order) => {
-                return monitoredOrder.id === order.id;
-              });
-              this.ordersToMonitor.splice(j, 1);
-            }
-          });
-
-          this.orderHistoryCache = _.reverse(_.sortBy(this.orderHistoryCache, 'orderDate'));
-          this.orderHistorySubject.next(this.orderHistoryCache);
-        },
-        error => console.error(error));
-  }
-
-  getEnvironmentDetails(): Promise<EnvironmentDetails> {
-    if (isDefined(this.environmentDetails)) {
-      return new Promise<EnvironmentDetails>(resolve => resolve(this.environmentDetails));
-    } else {
-      return this.fetchEnvironmentDetails();
-    }
-  }
-
   login(username: string, password: string): Promise<UserInfo> {
     return new Promise((resolve, reject) => {
-      this.http.post(this.config.backendUrl + 'login', {username: username, password: password})
+      this.http.post(environment.backendUrl + 'login', {username: username, password: password})
         .subscribe(
           (response: {sessionKey: string, userInfo: UserInfo}) => {
-                  this.sessionKey = response.sessionKey;
-                  this.user = response.userInfo;
-                  resolve(response.userInfo);
-                },
-          error => reject(error)
+            this.sessionKey = response.sessionKey;
+            this.user = response.userInfo;
+            this.userUpdates.next(this.user);
+            resolve();
+          },
+          error => reject('Invalid username or password') // todo: can do a better job of handling this
         );
     });
   }
@@ -118,12 +54,12 @@ export class MessageService implements OnDestroy {
   logout(): Promise<void> {
     return new Promise<void>((resolve, reject) => {
       const body = {username: this.user.username, sessionKey: this.sessionKey};
-      this.http.post(this.config.backendUrl + 'logout', body)
+      this.http.post(environment.backendUrl + 'logout', body)
         .subscribe(
           () => {
             this.sessionKey = undefined;
             this.user = undefined;
-            this.environmentDetails = undefined;
+            this.userUpdates.next(undefined);
             resolve();
           },
           error => reject(error)
@@ -138,36 +74,15 @@ export class MessageService implements OnDestroy {
         sessionKey: this.sessionKey,
         userInfo: user
       };
-      this.http.post(this.config.backendUrl + 'updateAccountInfo', body)
-        .subscribe(() => resolve(), error => reject(error));
+      this.http.post(environment.backendUrl + 'updateAccountInfo', body)
+        .subscribe(
+          () => {
+            this.user = user;
+            this.userUpdates.next(this.user);
+            resolve();
+          },
+          error => reject(error));
     });
-  }
-
-  getIncomingBatches(): Subject<Batch[]> {
-    if (!isDefined(this.batchUpdatesSubject)) {
-      this.batchUpdatesSubject = new Subject<Batch[]>();
-      const body = {username: this.user.username, sessionKey: this.sessionKey};
-      this.http.post(this.config.backendUrl + 'getIncomingBatches', body)
-        .subscribe((response: {batches: Batch[]}) => {
-          // send request now so that we don't have to wait the polling time till the first response
-          this.batchUpdatesCache = response.batches;
-          this.batchUpdatesSubject.next(this.batchUpdatesCache);
-        });
-
-      this.batchUpdatesTimer = Observable.interval(MessageService.BATCH_POLLING_RATE);
-      this.batchUpdatesSubscription = this.batchUpdatesTimer.subscribe(() => {
-        this.http.post(this.config.backendUrl + 'getIncomingBatches', body)
-          .subscribe((response: {batches: Batch[]}) => {
-            this.batchUpdatesCache = response.batches;
-            this.batchUpdatesSubject.next(this.batchUpdatesCache);
-          });
-      });
-
-      return this.batchUpdatesSubject;
-    } else {
-      setTimeout(() => this.batchUpdatesSubject.next(this.batchUpdatesCache));
-      return this.batchUpdatesSubject;
-    }
   }
 
   sendBatch(batch: Batch): Promise<void> {
@@ -179,39 +94,47 @@ export class MessageService implements OnDestroy {
       };
 
       // todo: can we just map this???
-      this.http.post(this.config.backendUrl + 'sendBatch', body).subscribe(() => resolve(), error => reject(error));
+      this.http.post(environment.backendUrl + 'sendBatch', body).subscribe(() => resolve(), error => reject(error));
     });
   }
 
-  getRobotStatusUpdates(): Promise<RobotStatus[]> {
-    return new Promise<RobotStatus[]>((resolve, reject) => {
-      const body = {
-        username: this.user.username,
-        sessionKey: this.sessionKey
-      };
+  getIncomingBatches(): Promise<Batch[]> {
+    return new Promise<Batch[]>((resolve, reject) => {
+      const body = {username: this.user.username, sessionKey: this.sessionKey};
+      this.http.post(environment.backendUrl + 'getIncomingBatches', body)
+        .subscribe(
+          (response: {batches: Batch[]}) => resolve(response.batches),
+          error => reject(error)
+        );
+    });
+  }
 
-      this.http.post(this.config.backendUrl + 'getRobotStatusUpdates', body).subscribe(
-        (response: {robotStatuses: RobotStatus[]}) => resolve(response.robotStatuses),
+  getOrderHistory(): Promise<Order[]> {
+    return new Promise<Order[]>((resolve, reject) => {
+      const body = {username: this.user.username, sessionKey: this.sessionKey};
+      this.http.post(environment.backendUrl + 'getOrderHistory', body).subscribe(
+        (response: {orderHistory: Order[]}) => resolve(response.orderHistory),
         error => reject(error)
       );
     });
   }
 
-  unsubscribeFromBatchUpdates(): void {
-    if (isDefined(this.batchUpdatesSubscription)) {
-      this.batchUpdatesSubscription.unsubscribe();
-      this.batchUpdatesSubscription = undefined;
-    }
-    if (isDefined(this.batchUpdatesTimer)) {
-      this.batchUpdatesTimer = undefined;
-    }
-    if (isDefined(this.batchUpdatesSubject)) {
-      this.batchUpdatesSubject.complete();
-      this.batchUpdatesSubject = undefined;
-    }
-
+  getOrderUpdates(ordersToMonitor: Order[]): Promise<Order[]> {
+    return new Promise<Order[]>((resolve, reject) => {
+      const body = {
+        username: this.user.username,
+        sessionKey: this.sessionKey,
+        orders: ordersToMonitor
+      };
+      this.http.post(environment.backendUrl + 'getOrderUpdates', body)
+        .subscribe(
+          (response: {orders: Order[]}) => resolve(response.orders),
+          error => reject(error)
+        );
+    });
   }
 
+  // todo: consolidate input arguments down
   placeOrder(selectedBeverage: OrderOption, selectedAddOns: {key: string, value: string | boolean | number}[],
              deliveryLocation: DeliveryLocation): Promise<Order> {
     return new Promise<Order>((resolve, reject) => {
@@ -221,82 +144,26 @@ export class MessageService implements OnDestroy {
         orderInfo: new OrderInfo(selectedBeverage, selectedAddOns),
         deliveryLocation: deliveryLocation
       };
-      this.http.post(this.config.backendUrl + 'placeOrder', body)
+      this.http.post(environment.backendUrl + 'placeOrder', body)
         .subscribe((response: {order: Order}) => {
-          if (response.order.state !== MessageService.DELIVERED) {
-            this.ordersToMonitor.push(response.order);
-            if (!isDefined(this.orderUpdatesSubscription)) {
-              this.orderUpdatesTimer = Observable.interval(MessageService.ORDER_POLLING_RATE);
-              this.orderUpdatesSubscription = this.orderUpdatesTimer.subscribe(() => this.handleGetOrderUpdates());
-            }
-          }
+          this.orderPlacedUpdate.next(response.order);
           resolve(response.order);
         }, error => reject(error));
     });
   }
 
-  getOrderHistory(): Subject<Order[]> {
-    if (!isDefined(this.orderHistorySubject)) {
-      this.orderHistorySubject = new Subject<Order[]>();
-      this.ordersToMonitor = [];
+  getSystemDetails(): Promise<SystemDetails> {
+    return new Promise<SystemDetails>((resolve, reject) => {
+      const body = {
+        username: this.user.username,
+        sessionKey: this.sessionKey
+      };
 
-      const body = {username: this.user.username, sessionKey: this.sessionKey};
-      this.http.post(this.config.backendUrl + 'getOrderHistory', body).subscribe(
-        (response: {orderHistory: Order}) => {
-                this.orderHistoryCache = _.reverse(_.sortBy(response.orderHistory, 'orderDate'));
-
-                this.orderHistoryCache.forEach((order: Order) => {
-                  if (order.state !== MessageService.DELIVERED) {
-                    this.ordersToMonitor.push(order);
-                  }
-                });
-                this.ordersToMonitor = _.uniqBy(this.ordersToMonitor, 'id');
-
-                if (this.ordersToMonitor.length !== 0 && !isDefined(this.orderUpdatesSubscription)) {
-                  this.orderUpdatesSubscription = Observable.interval(MessageService.ORDER_POLLING_RATE)
-                    .subscribe(() => this.handleGetOrderUpdates());
-                }
-
-                this.orderHistorySubject.next(this.orderHistoryCache);
-              },
-        error => console.error(error));
-      return this.orderHistorySubject;
-    }
-    setTimeout(() => this.orderHistorySubject.next(this.orderHistoryCache));
-    return this.orderHistorySubject;
-  }
-
-  unsubscribeFromOrderHistoryUpdates(): void {
-    if (isDefined(this.orderUpdatesSubscription)) {
-      this.orderUpdatesSubscription.unsubscribe();
-      this.orderUpdatesSubscription = undefined;
-    }
-    if (isDefined(this.orderUpdatesTimer)) {
-      this.orderUpdatesTimer = undefined;
-    }
-    if (isDefined(this.orderHistorySubject)) {
-      this.orderHistorySubject.complete();
-      this.orderHistorySubject = undefined;
-    }
-}
-
-  ngOnDestroy(): void {
-    if (isDefined(this.batchUpdatesSubscription)) {
-      this.batchUpdatesSubscription.unsubscribe();
-      this.batchUpdatesSubscription = undefined;
-    }
-    if (isDefined(this.batchUpdatesSubject)) {
-      this.batchUpdatesSubject.complete();
-      this.batchUpdatesSubject = undefined;
-    }
-    if (isDefined(this.orderUpdatesSubscription)) {
-      this.orderUpdatesSubscription.unsubscribe();
-      this.orderUpdatesSubscription = undefined;
-    }
-    if (isDefined(this.orderHistorySubject)) {
-      this.orderHistorySubject.complete();
-      this.orderHistorySubject = undefined;
-    }
+      this.http.post(environment.backendUrl + 'getSystemDetails', body).subscribe(
+        (response: SystemDetails) => resolve(response),
+        error => reject(error)
+      );
+    });
   }
 
 }