+++ src/environments/environment.prod.ts
@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  backendUrl: 'https://mobile-server-backend.herokuapp.com/ui/'
 };